Cambios en DAOs de Carrito, modificación global del funcionamiento del carrito

+++ src/main/java/curso/java/tienda/index/pojo/Producto.java
@@ -38,7 +38,7 @@ public class Producto {
 	@Column(name = "stock")
 	private int stock;
 	@Column(name = "impuesto")
-	private float impuesto;
+	private double impuesto;
 	@Column(name = "imagen")
 	private String imagen;
 	@Column(name = "baja", columnDefinition = "TINYINT", nullable = true)
